Rewrite of `inet_tcp_proxy_dist`
The previous version had the following issues:

* It crashed frequently, leading to failing testsuites in RabbitMQ,
  obviously for reasons unrelated to the tests themselves. This even
  affected the communication between the common_test control node and
  the RabbitMQ nodes.

* It was inefficient. For instance, it took 60 seconds to cluster 5
  RabbitMQ nodes (not counting the time to start them initially).

The new implementation is incompatible with the previous one. I took the
skeletton of `inet_tcp_compress_dist` (closed source as of this commit).
The main differences are:

* The new implementation does not use a new set of TCP ports and does
  not manage a different TCP connection. The connection started initially
  is used. It manages an internal state, using `persistent_term` instead
  of an ETS table, to record which peers are blocked.

* Because the existing TCP connection is used, it uses internal buffers
  to keep input and output data until the connection is allowed. When
  the connection is allowed again, those buffers are flushed.

This dist module should behave the same as the regular default
distribution module when connections are allowed.

The build system changed from Rebar 3 to Erlang.mk in the process.