inet_tcp_proxy_dist: Do not block on the input code path
When the connection between the local node and the remote node is
blocked, we now only block the output code path (sending from the local
to the remote node).

The input code path (receiving from the remote node) is left alone. The
goal is to be able to simulate asymmetrical communication.

The input buffering is removed at the same time because we don't block
anymore.